Move `import_blocks` out of `dsn` submodule and make it private
